Rename WalkWatcher -> Watcher to match naming pattern
