+++ tests/watcherconfig_test.py
@@ -1,8 +1,13 @@
 from __future__ import annotations
 
+import os
+import tempfile
+
 import pytest
 
+from walk_watcher.watcherconfig import NEW_CONFIG
 from walk_watcher.watcherconfig import WatcherConfig
+from walk_watcher.watcherconfig import write_new_config
 
 CONFIG_PATH = "tests/test_config.ini"
 
@@ -27,3 +32,37 @@ def test_watcherconfig_loads_test_fixture_completely() -> None:
 
     assert config.exclude_directory_pattern == r"\/directory02|fixture$|\\directory02"
     assert config.exclude_file_pattern == "file01.*"
+
+
+def test_write_new_config() -> None:
+    try:
+        fd, filename = tempfile.mkstemp(suffix=".ini")
+        os.close(fd)
+        os.remove(filename)
+        expected = NEW_CONFIG.format(filename=filename.replace(".ini", ".db"))
+
+        write_new_config(filename)
+
+        with open(filename) as f:
+            content = f.read()
+
+        assert content == expected
+
+    finally:
+        os.remove(filename)
+
+
+def test_write_new_config_early_exit_when_exists() -> None:
+    try:
+        fd, filename = tempfile.mkstemp(suffix=".ini")
+        os.close(fd)
+
+        write_new_config(filename)
+
+        with open(filename) as f:
+            content = f.read()
+
+        assert content == ""
+
+    finally:
+        os.remove(filename)
